Add operators and loaders
